backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
pydantic==1.10.13          # SQLModel пока использует Pydantic v1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
bcrypt>=4.1.3          # самая свежая версия
